+++ AnimeTrackerApplication/Logic/Animes/IAnime.cs
@@ -13,6 +13,6 @@ namespace Logic.Animes
         void DeleteAnime(string animeId);
         List<Anime> GetAllAnime();
         int GetNrAnime();
-        void UpdateAnime(string id, string name, string studio, int nrEpisodes, int releaseYear, Season releaseSeason, double rating, string description, string imageURL);
+        void UpdateAnime(string id, string name, string studio, int nrEpisodes, int releaseYear, Season releaseSeason, decimal rating, string description, string imageURL);
     }
 }
